requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
discord.py @ git+https://github.com/Rapptz/discord.py@a106201ad394c207df222521842de796a460f27f
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
yarl==1.7.2
